X11Adapter: Rework async X11 handling
Add CHECK() macro for only checking for errors. Provide overloads for
either calling a specific callback or just log something and optionally
fire a generic event on a given id.

Document usage of AWAIT() and CHECK() macros.

Improve logic for injecting a "sync" request to only do that when really
necessary.

Adapt Window, Font and TextRenderer to minimize requests and only use
CHECK() instead of AWAIT() for requests without a reply.