TextRenderer: Add glitch mode for compositeglyphs
At least with CompositeGlyphs XRender requests, and at least on my
Radeon R600 when using Glamor instead of EXA, the origin for the src
picture is wrong. This can be "fixed" by adding a picture
transformation, which will instead break rendering on systems without
this bug. Therefore make it optional.

For now, a glitch flag in X11Adapter can be set from the commandline
with "-glitches 1". Not sure whether it's possible to detect when a
workaround is needed...