X11App: Unify last-resort error handling here
Provide a mechanism to raise an error event with a Widget and a Window
attached. Adapt all other modules to use this for anything that isn't
explicitly handled. The event args provide a way to ignore the error. If
this isn't done, the details (including the Widget's class and name as
well as the Window name) are logged and the application is terminated.