Drop Timer, use PSC_Timer instead
Moved timer implementation to poser-core, where it can be integrated
with the event loop. This way, it can use notification by signals
instead of the clunky thread notification interface, also eliminating
the need for a dummy pipe to get the notification back to the main
thread. Furthermore, timers are now multiplexed, so the system limit of
timers (max 32) only applies to timers concurrently running, not total
instances.