Window: Fix minimizing with kwin rejecting focus
KDE's kwin forbids minimizing a window that does not accept any input.
This can be avoided by adding a (bogus) WM_TAKE_FOCUS to WM_PROTOCOLS
which signals we *might* accept focus and causes kwin to allow
minimizing the window again.

We'll still ignore any focus offer and use a keyboard grab to allow
input in the search field, so focus reverts back to the last focused
window automatically when the grab is released.