X11Adapter: Add code probing for rendering glitch
Add code compositing a dummy glyph to a 1px offscreen pixmap with an
y offset of 1 pixel. This should result in nothing. If the one pixel is
set nevertheless, we know we're on a system with buggy glyph compositing
behavior and can enable the workaround automatically.

Drop the "glitches" resource / command-line setting, not needed any
more.