Font: Accept any result for last fallback
It seems we can't rely on fontconfig to always include the family name
we asked for when an exact match was found, e.g. the fontconfig version
from OpenBSD returns "Noto Color Emoji" when asked for "emoji", but
doesn't list "emoji" in the family name property of the result.

Therefore, just accept anything from fontconfig for the last fallback
search pattern listed.