X11App: Fix checking locale on NetBSD
On NetBSD, setlocale(LC_ALL, "") returns a string with multiple entries
separated by / instead of just the single LC_ALL value that was set. It
seems the last of these entries is LC_ALL, so when getting a string
containing slashes, assume we have to look at the last entry.