X11App: Fix enforcing UTF-8 locale
To enforce UTF-8 encoding, only the LC_CTYPE category is relevant. So,
if we get a string with multiple entries like on NetBSD, only look at
the one for LC_CTYPE, which seems to be the second one. Also, when
attempting to modify the locale, only modify LC_CTYPE in any case.