X11Adapter: Document glyph rendering glitch
Document the glyph rendering glitch X11Adapter probes for and also the
actual probing method. Factor it out to a separate function.

The glitch affects XRender CompositeGlyphs* requests; when it occurs,
the coordinates of the source picture are ignored and the source picture
is always aligned to the origin of the destination picture and also
repeated in both dimensions. My suspicion is that it is caused by a bug
in the glamor implementation of XRender.